requirements.txt
Django==3.1.1
django-cors-headers==3.2.1
djangorestframework==3.11.0
Pillow==6.2.1
django-ckeditor==6.0.0
django-rest-knox==4.1.0